Corrections en cours.

+++ controllers/sauce.js
@@ -125,19 +125,28 @@ exports.deleteSauce = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
     // Dans le cas où la requête contient un fichier, on cherche l'imageUrl de la sauce existante et on supprime l'ancien fichier du dossier images
     if (req.file) {
+        //console.log(req.file);
         Sauce.findOne({_id: req.params.id}, (e,sauce) => {
             if (e) {
                 console.error("La recherche de l'imageUrl d'origine a échoué...");
             }
             else {
                 const fileToDelete = sauce.imageUrl.split('/images/')[1];
-                fs.unlink(`images/${fileToDelete}`, (e) => {
-                    if (e) {console.error('Echec de la suppression de l\'ancien fichier');}
-                    else {console.log('Ancienne image supprimée avec succès.');}
-                })
+                if (fileToDelete === req.file.filename) {
+                    console.log("*** " + fileToDelete + " ***");
+                    console.log("***** " + req.file.filename + " *****");
+                }
+                else {
+                    //console.log(fileToDelete);
+                    fs.unlink(`images/${fileToDelete}`, (e) => {
+                        if (e) {console.error(/* 'Echec de la suppression de l\'ancien fichier' */e);}
+                        else {console.log('Ancienne image supprimée avec succès.');}
+                    })
+                }
             }
         })
     }
+    
     const sauceObject = req.file ?
     // S'il y a une image dans la requête en remplacement de celle existante dans le BDD :
         {
